requirements.txt
transformers>=4.40.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.5
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0
pypdf>=3.17.0
typing-extensions>=4.7.1
langgraph>=0.0.20
streamlit>=1.35.0